CUDA collation with MPI type

+++ src/cuda/collated_cudaprofiling_types.h
@@ -8,7 +8,7 @@ typedef struct {
    int cbid;
    int n_calls;
    float t_ms;
-   size_t memcpy_bytes[2];
+   long long memcpy_bytes[2];
    long long overhead_nsec;
 } collated_cudaprofile_t;
 #endif
