+++ src/cuda/cudaprofiling.c
@@ -22,7 +22,7 @@ cudaprofile_t vftr_new_cudaprofiling() {
 }
 
 void vftr_accumulate_cudaprofiling (cudaprofile_t *prof, int cbid, int n_calls,
-                                    float t_ms, int mem_dir, uint64_t memcpy_bytes) {
+                                    float t_ms, int mem_dir, long long memcpy_bytes) {
    prof->cbid = cbid;
    prof->n_calls += n_calls;
    prof->t_ms += t_ms;
