+++ src/cuda/cudaprofiling_types.h
@@ -11,7 +11,7 @@ typedef struct {
    int cbid;
    int n_calls;
    float t_ms;
-   size_t memcpy_bytes[2];
+   long long memcpy_bytes[2];
    long long overhead_nsec;
 } cudaprofile_t;
 #endif
